+++ packages/terra-date-time-picker/src/DateTimePicker.jsx
@@ -99,7 +99,9 @@ const propTypes = {
    */
   value: PropTypes.string,
   /**
-   * Type of time input to initialize. Must be `24-hour` or `12-hour`
+   * Type of time input to initialize. Must be `24-hour` or `12-hour`.
+   * The `de`, `es-ES`, `fr-FR`, `fr`, `nl-BE`, `nl`, `pt-BR`, `pt`, `sv-SE` and `sv` locales do not use the 12-hour time notation.
+   * If the `variant` prop if set to `12-hour` for one of these supported locales, the variant will be ignored and defaults to `24-hour`.
    */
   timeVariant: PropTypes.oneOf([DateTimeUtils.FORMAT_12_HOUR, DateTimeUtils.FORMAT_24_HOUR]),
 };