+++ packages/terra-layout/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+4.14.0 - (October 3, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 4.13.0 - (September 26, 2019)
 ------------------
 ### Changed