+++ packages/terra-time-input/CHANGELOG.md
@@ -7,6 +7,11 @@ Unreleased
 * Changed 12-hour format appearance for mobile and desktop.
   * Meridiem select on Desktop has been changed to buttons which means that the WebdriverIO reference screenshots (if any) and/or any code in the consumer application that accesses the old meridiem select functionality on the Desktop view need to be updated.
 
+4.12.0 - (October 3, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 4.11.0 - (September 26, 2019)
 ------------------
 ### Changed
@@ -16,6 +21,7 @@ Unreleased
 ------------------
 * Removed `details` tag from doc-site.
 * Doc site example module import syntax modified to use package.json name resolution
+* Fallback the time notation variant to `24-hour` for locales that do not use the `12-hour` notation.
 
 ### Added
 * Added en-CA translations.